build: Pin Composer requirement versions
It looks like the getopt-php package has changed slightly
to now have the --help message end with an extra new line,
which is causing the tests to fail.

This change is fine, and we can update the tests, but we should
not get these failures in the future unless we intentionally
update the requirements.